+++ packages/presets/client/CHANGELOG.md
@@ -1,5 +1,81 @@
 # @graphql-codegen/client-preset
 
+## 4.3.2
+
+### Patch Changes
+
+- Updated dependencies [[`79fee3c`](https://github.com/dotansimha/graphql-code-generator/commit/79fee3cada20d683d250aad5aa5fef9d6ed9f4d2)]:
+  - @graphql-codegen/visitor-plugin-common@5.3.1
+  - @graphql-codegen/gql-tag-operations@4.0.9
+  - @graphql-codegen/typescript-operations@4.2.3
+  - @graphql-codegen/typed-document-node@5.0.9
+  - @graphql-codegen/typescript@4.0.9
+
+## 4.3.1
+
+### Patch Changes
+
+- Updated dependencies [[`808ada5`](https://github.com/dotansimha/graphql-code-generator/commit/808ada595d83d39cad045da5824cac6378e9eca3), [`14ce39e`](https://github.com/dotansimha/graphql-code-generator/commit/14ce39e41dfee38c652be736664177fa2b1df421)]:
+  - @graphql-codegen/visitor-plugin-common@5.3.0
+  - @graphql-codegen/gql-tag-operations@4.0.8
+  - @graphql-codegen/typescript-operations@4.2.2
+  - @graphql-codegen/typed-document-node@5.0.8
+  - @graphql-codegen/typescript@4.0.8
+
+## 4.3.0
+
+### Minor Changes
+
+- [#10001](https://github.com/dotansimha/graphql-code-generator/pull/10001) [`1be6e65`](https://github.com/dotansimha/graphql-code-generator/commit/1be6e65943b85162f3d465189d0a6df4b962df5d) Thanks [@n1ru4l](https://github.com/n1ru4l)! - Support discriminating `null` and `undefined` within the `useFragment` function.
+
+  ```ts
+  function MyComponent(props: FragmentType<typeof MyFragment> | null) {
+    const data = useFragment(MyFragment, props);
+    // data is `MyFragment | null`
+  }
+
+  function MyComponent(props: FragmentType<typeof MyFragment> | undefined) {
+    const data = useFragment(MyFragment, props);
+    // data is `MyFragment | undefined`
+  }
+  ```
+
+  Before, the returned type from `useFragment` was always `TType | null | undefined`.
+
+- [#9804](https://github.com/dotansimha/graphql-code-generator/pull/9804) [`5e594ef`](https://github.com/dotansimha/graphql-code-generator/commit/5e594ef8f39b9e1036b6bcaa977f914a66fec03e) Thanks [@rachel-church](https://github.com/rachel-church)! - Preserving `Array<T>` or `ReadonlyArray<T>` in `useFragment()` return type.
+
+### Patch Changes
+
+- [#9996](https://github.com/dotansimha/graphql-code-generator/pull/9996) [`99f449c`](https://github.com/dotansimha/graphql-code-generator/commit/99f449c8dcd645d49eda26e4ddfcb8ad7056ecbf) Thanks [@nahn20](https://github.com/nahn20)! - Added configuration to allow for custom hash functions for persisted documents in the client preset
+
+  ### Example
+
+  ```ts filename="codegen.ts" {10-12}
+  import { type CodegenConfig } from '@graphql-codegen/cli';
+
+  const config: CodegenConfig = {
+    schema: 'schema.graphql',
+    documents: ['src/**/*.tsx'],
+    generates: {
+      './src/gql/': {
+        preset: 'client',
+        presetConfig: {
+          persistedDocuments: {
+            hashAlgorithm: operation => {
+              const shasum = crypto.createHash('sha512');
+              shasum.update(operation);
+              return shasum.digest('hex');
+            },
+          },
+        },
+      },
+    },
+  };
+  ```
+
+- Updated dependencies [[`5501c62`](https://github.com/dotansimha/graphql-code-generator/commit/5501c621f19eb5ef8e703a21f7367e07e41f199c)]:
+  - @graphql-codegen/add@5.0.3
+
 ## 4.2.6
 
 ### Patch Changes